refactor: delete AdminLoginController and updated packages and test
