+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:41913/jest?","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:35239/jest?","mongoDBName":"jest"}
