+++ package.json
@@ -36,9 +36,6 @@
     "mongoose": "^5.10.6",
     "password-generator": "^2.3.2",
     "path": "^0.12.7",
-    "pg": "^8.3.3",
-    "pg-hstore": "^2.3.3",
-    "sequelize": "^6.3.5",
     "uuid": "^8.3.0"
   },
   "jest": {
