--- src/controllers/AdminLoginController.js
@@ -1,26 +1,0 @@
-const Admin = require('../models/Admin');
-
-module.exports = {
-
-    async show(req, res) {
-        try {
-            const { email, password } = req.body;
-
-            const user = await Admin.findOne({ email });
-
-            if (user != null) {
-                if (user.password === password) {
-                    return res.status(201).json(user);
-                }
-                if (user.password !== password) {
-                    return res.status(400).json('Senha Incorreta');
-                }
-            }
-
-            return res.status(404).json('Usuário não encontrado');
-        } catch (err) {
-            return res.status(500).json({ error: err.message });
-        }
-    },
-
-};
